+++ src/main/kotlin/net/rebux/jumpandrun/events/ParkourFinishEvent.kt
@@ -7,7 +7,7 @@ import org.bukkit.event.HandlerList
 /**
  * An [Event] that is being called when a [player] finishes a parkour
  */
-class ParkourFinishEvent(val player: Player, val ticks: Int) : Event() {
+class ParkourFinishEvent(val player: Player, val ticks: Long) : Event() {
 
     @Override
     override fun getHandlers(): HandlerList {
